backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.6.0
pydantic[email]==2.10.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1
httpx==0.28.0
Pillow==11.1.0
numpy==2.2.0
groq>=0.11.0
scikit-learn==1.6.0
joblib==1.4.2
pandas==2.2.3
email-validator==2.2.0
pytest==8.3.0
pytest-asyncio==0.24.0
certifi>=2024.2.2
